Added timeout for critical commands
There can be a case where commands such as ceph mon getmap can take
sometime but this being an async call the code would move forward which
can lead to a condition where the monmap being saved is corrupt. Hence
we wait for some time for the command to complete. If there is a timeout
we try to get monmap from other mon server.

Also added sleep time before check commands after every rectification
step as the rectificaton may take some time to reflect in the ceph
health command