Added support when ceph cli is down initially
Added support for the case when ceph cli is down initially i.e.
the quorum is not being established. In this scenario we try to
start ceph mon service at each of the reachable node. If that
does not work we start the checklist for the ceph_cli path i.e.
try to check for clock skew, look at the monmap issues